Hoist up NodeDisplayData to WorkflowMetaDisplay
