+++ ee/vellum_ee/workflows/display/vellum.py
@@ -1,12 +1,7 @@
 from dataclasses import dataclass
 from uuid import UUID
-from typing import List, Literal, Optional, Union
+from typing import List, Literal, Optional
 
-from pydantic import Field
-
-from vellum import VellumVariableType
-from vellum.client.types.array_vellum_value import ArrayVellumValue
-from vellum.client.types.vellum_value import VellumValue
 from vellum.core import UniversalBaseModel
 from vellum_ee.workflows.display.base import (
     EdgeDisplay,
@@ -22,23 +17,10 @@ from vellum_ee.workflows.display.base import (
 )
 from vellum_ee.workflows.display.base import WorkflowDisplayData  # noqa: F401 - Remove in 0.15.0
 from vellum_ee.workflows.display.base import WorkflowDisplayDataViewport  # noqa: F401 - Remove in 0.15.0
-
-
-class NodeDisplayPosition(UniversalBaseModel):
-    x: float = 0.0
-    y: float = 0.0
-
-
-class NodeDisplayComment(UniversalBaseModel):
-    value: Optional[str] = None
-    expanded: Optional[bool] = None
-
-
-class NodeDisplayData(UniversalBaseModel):
-    position: NodeDisplayPosition = Field(default_factory=NodeDisplayPosition)
-    width: Optional[int] = None
-    height: Optional[int] = None
-    comment: Optional[NodeDisplayComment] = None
+from vellum_ee.workflows.display.editor.types import NodeDisplayComment  # noqa: F401 - Remove in 0.15.0
+from vellum_ee.workflows.display.editor.types import NodeDisplayData
+from vellum_ee.workflows.display.editor.types import NodeDisplayPosition  # noqa: F401 - Remove in 0.15.0
+from vellum_ee.workflows.display.utils.vellum import NodeInputValuePointerRule
 
 
 class CodeResourceDefinition(UniversalBaseModel):
@@ -48,7 +30,11 @@ class CodeResourceDefinition(UniversalBaseModel):
 
 @dataclass
 class WorkflowMetaVellumDisplayOverrides(WorkflowMetaDisplayOverrides):
-    entrypoint_node_display: NodeDisplayData = Field(default_factory=NodeDisplayData)
+    """
+    DEPRECATED: Use WorkflowMetaDisplay instead. Will be removed in 0.15.0
+    """
+
+    pass
 
 
 @dataclass
@@ -135,61 +121,6 @@ class WorkflowOutputVellumDisplay(WorkflowOutputVellumDisplayOverrides):
     """
 
     pass
-
-
-class ConstantValuePointer(UniversalBaseModel):
-    type: Literal["CONSTANT_VALUE"] = "CONSTANT_VALUE"
-    data: VellumValue
-
-
-ArrayVellumValue.model_rebuild()
-
-
-class NodeOutputData(UniversalBaseModel):
-    node_id: str
-    output_id: str
-
-
-class NodeOutputPointer(UniversalBaseModel):
-    type: Literal["NODE_OUTPUT"] = "NODE_OUTPUT"
-    data: NodeOutputData
-
-
-class InputVariableData(UniversalBaseModel):
-    input_variable_id: str
-
-
-class InputVariablePointer(UniversalBaseModel):
-    type: Literal["INPUT_VARIABLE"] = "INPUT_VARIABLE"
-    data: InputVariableData
-
-
-class WorkspaceSecretData(UniversalBaseModel):
-    type: VellumVariableType
-    workspace_secret_id: Optional[str] = None
-
-
-class WorkspaceSecretPointer(UniversalBaseModel):
-    type: Literal["WORKSPACE_SECRET"] = "WORKSPACE_SECRET"
-    data: WorkspaceSecretData
-
-
-class ExecutionCounterData(UniversalBaseModel):
-    node_id: str
-
-
-class ExecutionCounterPointer(UniversalBaseModel):
-    type: Literal["EXECUTION_COUNTER"] = "EXECUTION_COUNTER"
-    data: ExecutionCounterData
-
-
-NodeInputValuePointerRule = Union[
-    NodeOutputPointer,
-    InputVariablePointer,
-    ConstantValuePointer,
-    WorkspaceSecretPointer,
-    ExecutionCounterPointer,
-]
 
 
 class NodeInputValuePointer(UniversalBaseModel):
